SFEperl-date-manip: bump to 6.58, package for perl-522 is added, for perl-520 is obsolete
--HG--
branch : spec_SFEperl-date-manip

+++ specs/SFEperl-date-manip.spec
@@ -4,7 +4,7 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 1
 
 %define cpan_name Date-Manip
@@ -14,8 +14,8 @@
 Summary:               Date manipulation routines
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               6.52
-IPS_component_version: 6.52
+Version:               6.58
+IPS_component_version: 6.58
 License:               perl_5
 URL:                   https://metacpan.org/pod/Date::Manip
 Source0:               http://cpan.metacpan.org/authors/id/S/SB/SBECK/Date-Manip-%{version}.tar.gz
@@ -32,6 +32,12 @@ BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/test-inter-584
 BuildRequires:    library/perl-5/test-simple-584
+BuildRequires:    library/perl-5/carp-584
+BuildRequires:    library/perl-5/data-dumper-584
+BuildRequires:    library/perl-5/encode-584
+BuildRequires:    library/perl-5/io-584
+BuildRequires:    library/perl-5/pathtools-584
+BuildRequires:    library/perl-5/storable-584
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-584
@@ -53,6 +59,12 @@ BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
 BuildRequires:    library/perl-5/test-inter-510
 BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/carp-510
+BuildRequires:    library/perl-5/data-dumper-510
+BuildRequires:    library/perl-5/encode-510
+BuildRequires:    library/perl-5/io-510
+BuildRequires:    library/perl-5/pathtools-510
+BuildRequires:    library/perl-5/storable-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-510
@@ -74,6 +86,12 @@ BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/test-inter-512
 BuildRequires:    library/perl-5/test-simple-512
+BuildRequires:    library/perl-5/carp-512
+BuildRequires:    library/perl-5/data-dumper-512
+BuildRequires:    library/perl-5/encode-512
+BuildRequires:    library/perl-5/io-512
+BuildRequires:    library/perl-5/pathtools-512
+BuildRequires:    library/perl-5/storable-512
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-512
@@ -95,6 +113,13 @@ BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/test-inter-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/carp-516
+BuildRequires:    library/perl-5/data-dumper-516
+BuildRequires:    library/perl-5/encode-516
+BuildRequires:    library/perl-5/io-516
+BuildRequires:    library/perl-5/pathtools-516
+BuildRequires:    library/perl-5/storable-516
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-516
@@ -108,31 +133,37 @@ Requires:         library/perl-5/storable-516
 Date manipulation routines
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Date manipulation routines
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/test-inter-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-inter-522
+BuildRequires:    library/perl-5/test-simple-522
+BuildRequires:    library/perl-5/carp-522
+BuildRequires:    library/perl-5/data-dumper-522
+BuildRequires:    library/perl-5/encode-522
+BuildRequires:    library/perl-5/io-522
+BuildRequires:    library/perl-5/pathtools-522
+BuildRequires:    library/perl-5/storable-522
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/carp-520
-Requires:         library/perl-5/data-dumper-520
-Requires:         library/perl-5/encode-520
-Requires:         library/perl-5/io-520
-Requires:         library/perl-5/pathtools-520
-Requires:         library/perl-5/storable-520
+Requires:         library/perl-5/carp-522
+Requires:         library/perl-5/data-dumper-522
+Requires:         library/perl-5/encode-522
+Requires:         library/perl-5/io-522
+Requires:         library/perl-5/pathtools-522
+Requires:         library/perl-5/storable-522
 
-%description 520
+%description 522
 Date manipulation routines
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -145,8 +176,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -161,7 +201,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -235,8 +275,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -297,18 +337,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 05 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 6.58, package for perl-522 is added, for perl-520 is obsolete
 * Sat Dec 05 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 6.52
 * Thu Nov 14 2013 Fumihisa TONAKA <fumi.ftnk@gmail.com>
